Update the algorithm to handle roles where naming is not supported
Add a statement to the end of Step 1 ("initialize") stating that if the
root node's role doesn't support naming, return the empty string. At the
present time, this impacts nothing because all roles currently support
naming. If that changes within the ARIA spec, this modification is meant
to accomplish the following:

<span aria-label="bar">foo</span>: The span has no name.

<button><span aria-label="bar">foo</span>: The button's accessible name
is "bar" (but the span itself has no name).

In other words, it prohibits generic elements themselves from getting
names, but preserves the name calculation results authors might be
expecting in currently-deployed content.